+++ src/api/m3u_api.rs
@@ -1,10 +1,11 @@
 use actix_web::{HttpRequest, HttpResponse, web};
-use log::error;
+use log::{error};
 use futures::{stream};
 use bytes::Bytes;
 
 use crate::api::api_utils::{get_user_target, get_user_target_by_credentials, stream_response};
 use crate::api::api_model::{AppState, UserApiRequest};
+use crate::model::api_proxy::ProxyType;
 use crate::model::config::TargetType;
 use crate::repository::m3u_repository::{m3u_get_file_paths, m3u_get_item_for_stream_id, m3u_load_rewrite_playlist};
 use crate::repository::storage::get_target_storage_path;
@@ -54,13 +55,18 @@ async fn m3u_api_stream(
 ) -> HttpResponse {
     let (username, password, stream_id) = path.into_inner();
     if let Ok(m3u_stream_id) = stream_id.parse::<u32>() {
-        if let Some((_user, target)) = get_user_target_by_credentials(&username, &password, &api_req, &app_state) {
+        if let Some((user, target)) = get_user_target_by_credentials(&username, &password, &api_req, &app_state) {
             if target.has_output(&TargetType::M3u) {
                 match get_target_storage_path(&app_state.config, target.name.as_str()) {
                     Some(target_path) => {
                         let (m3u_path, idx_path) = m3u_get_file_paths(&target_path);
                         match m3u_get_item_for_stream_id(&app_state.config, m3u_stream_id, &m3u_path, &idx_path) {
                             Ok(m3u_item) => {
+                                if user.proxy == ProxyType::Redirect {
+                                    let stream_url = m3u_item.url;
+                                    // debug!("Redirecting stream request to {stream_url}");
+                                    return HttpResponse::Found().insert_header(("Location", stream_url.to_string())).finish();
+                                }
                                 return stream_response(m3u_item.url.as_str(), &req, None).await;
                             }
                             Err(err) => {