+++ src/model/config.rs
@@ -304,7 +304,9 @@ pub struct ConfigTargetOptions {
     #[serde(default = "default_as_two_u16")]
     pub xtream_resolve_series_delay: u16,
     #[serde(default)]
-    pub m3u_use_type_url: bool,
+    pub m3u_include_type_in_url: bool,
+    #[serde(default)]
+    pub m3u_mask_redirect_url: bool,
 }
 
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]